requirements.txt
streamlit
pandas
langchain-openai
crewai
python-pptx
PyPDF2
pydantic